+++ src/bc5.rs
@@ -23,13 +23,13 @@ pub fn compress_blocks_into(surface: &RgbaSurface, blocks: &mut [u8]) {
 
     let mut rg_data = vec![0_u8; (surface.width * surface.height * 2) as usize];
     let pitch = (surface.width * 32 + 7) / 8;
-    let mut offset = 0_u32;
+    let mut offset = 0;
 
     for y in 0..surface.height {
         for x in 0..surface.width {
             // Copy R and G bytes over
-            rg_data[offset as usize] = surface.data[(x * 4 + y * pitch) as usize];
-            rg_data[(offset + 1) as usize] = surface.data[(x * 4 + y * pitch + 1) as usize];
+            rg_data[offset] = surface.data[(x * 4 + y * pitch) as usize];
+            rg_data[offset + 1] = surface.data[(x * 4 + y * pitch + 1) as usize];
             offset += 2;
         }
     }
